Reorganize projects and redo namespaces
