+++ src/MEU.GV4.Base/Models/Character.cs
@@ -1,4 +1,4 @@
-﻿namespace MEU.GV4.Data.Models;
+﻿namespace MEU.GV4.Base.Models;
 
 /// <summary>
 /// The Character class is the base class that all other character types inherit from
