+++ src/MEU.GV4.Base/Models/Game.cs
@@ -1,4 +1,4 @@
-﻿namespace MEU.GV4.Data.Models;
+﻿namespace MEU.GV4.Base.Models;
 
 /// <summary>
 /// The Game class is the top-level container for all game-related data.
