+++ src/MEU.GV4.Base/Models/Trait.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MEU.GV4.Data.Models;
+namespace MEU.GV4.Base.Models;
 
 public class Trait
 {
