+++ src/MEU.GV4.Base/Models/TraitList.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MEU.GV4.Data.Models;
+namespace MEU.GV4.Base.Models;
 
 public class TraitList : List<Trait>
 {
