+++ src/MEU.GV4.Base/Properties/AssemblyInfo.cs
@@ -12,7 +12,7 @@ using System.Runtime.InteropServices;
 // attribute to true on that type.
 
 [assembly: ComVisible(false)]
-[assembly:InternalsVisibleTo("MEU.GV4.Data.Tests")]
+[assembly:InternalsVisibleTo("MEU.GV4.Tests")]
 // The following GUID is for the ID of the typelib if this project is exposed to COM.
 
 [assembly: Guid("950b600b-a665-48f4-85db-9dc772cc111d")]
