+++ src/MEU.GV4.METClassic/Serialization/METTypeResolver.cs
@@ -1,4 +1,4 @@
-﻿using MEU.GV4.Data.Models;
+﻿using MEU.GV4.Base.Models;
 using MEU.GV4.METClassic.Models;
 using System.Reflection;
 using System.Text.Json;
