+++ src/MEU.GV4.Tests/MEU.GV4.Tests.csproj
@@ -15,8 +15,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\MEU.GV4.Data\MEU.GV4.Data.csproj" />
-    <ProjectReference Include="..\..\src\MEU.GV4.METClassic\MEU.GV4.METClassic.csproj" />
+    <ProjectReference Include="..\MEU.GV4.Base\MEU.GV4.Base.csproj" />
+    <ProjectReference Include="..\MEU.GV4.METClassic\MEU.GV4.METClassic.csproj" />
   </ItemGroup>
 
   <ItemGroup>
