+++ src/MEU.GV4.Tests/Serialization/GrapevineLegacyXMLReaderCharacterTests.cs
@@ -1,9 +1,9 @@
-﻿using MEU.GV4.Data.Models;
+﻿using MEU.GV4.Base.Models;
 using MEU.GV4.METClassic.Models;
 using MEU.GV4.METClassic.Serialization;
 using System.Xml.Linq;
 
-namespace MEU.GV4.Data.Tests.Providers;
+namespace MEU.GV4.Tests.Serialization;
 
 public class GrapevineLegacyXMLReaderCharacterTests
 {
